refactor(releases): moving ReleasePublishAllButton to separate component
