+++ packages/sanity/src/core/releases/tool/detail/ReleaseDetail.tsx
@@ -1,15 +1,15 @@
-import {ArrowLeftIcon, ErrorOutlineIcon, PublishIcon} from '@sanity/icons'
-import {Box, Card, Container, Flex, Heading, Stack, Text, useToast} from '@sanity/ui'
-import {useCallback, useEffect, useMemo, useState} from 'react'
+import {ArrowLeftIcon} from '@sanity/icons'
+import {Box, Card, Container, Flex, Heading, Stack, Text} from '@sanity/ui'
+import {useCallback, useEffect, useMemo} from 'react'
 import {LoadingBlock, useClient} from 'sanity'
 import {type RouterContextValue, useRouter} from 'sanity/router'
 
-import {Button, Dialog} from '../../../../ui-components'
+import {Button} from '../../../../ui-components'
 import {useListener} from '../../../hooks/useListener'
 import {useBundles} from '../../../store/bundles'
-import {useBundleOperations} from '../../../store/bundles/useBundleOperations'
 import {API_VERSION} from '../../../tasks/constants'
 import {BundleMenuButton} from '../../components/BundleMenuButton/BundleMenuButton'
+import {ReleasePublishAllButton} from '../../components/ReleasePublishAllButton/ReleasePublishAllButton'
 import {type ReleasesRouterState} from '../../types/router'
 import {useReleaseHistory} from './documentTable/useReleaseHistory'
 import {ReleaseReview} from './ReleaseReview'
@@ -44,29 +44,15 @@ export const ReleaseDetail = () => {
   const {bundleSlug}: ReleasesRouterState = router.state
   const parsedSlug = decodeURIComponent(bundleSlug || '')
   const {data, loading} = useBundles()
-  const {publishBundle} = useBundleOperations()
   const {documents: bundleDocuments, loading: documentsLoading} =
     useFetchBundleDocuments(parsedSlug)
-  const [publishBundleStatus, setPublishBundleStatus] = useState<'idle' | 'confirm' | 'publishing'>(
-    'idle',
-  )
   const history = useReleaseHistory(bundleDocuments)
   const validation = useBundleDocumentsValidation(bundleDocuments)
 
   const bundle = data?.find((storeBundle) => storeBundle.slug === parsedSlug)
   const bundleHasDocuments = !!bundleDocuments.length
   const showPublishButton = loading || !bundle?.publishedAt
-  const isValidatingDocuments = Object.values(validation).some(({isValidating}) => isValidating)
-  const hasDocumentValidationErrors = Object.values(validation).some(({hasError}) => hasError)
-  const isPublishButtonDisabled =
-    loading ||
-    !bundle ||
-    !bundleHasDocuments ||
-    publishBundleStatus === 'publishing' ||
-    hasDocumentValidationErrors ||
-    isValidatingDocuments
-
-  const toast = useToast()
+  const isPublishButtonDisabled = loading || !bundle || !bundleHasDocuments
 
   const navigateToReview = useCallback(() => {
     router.navigate({
@@ -82,31 +68,6 @@ export const ReleaseDetail = () => {
     })
   }, [router])
 
-  const publishTooltipContent = useMemo(() => {
-    if (!hasDocumentValidationErrors && !isValidatingDocuments) return null
-
-    const tooltipText = () => {
-      if (isValidatingDocuments) {
-        return 'Validating documents...'
-      }
-
-      if (hasDocumentValidationErrors) {
-        return 'Some documents have validation errors'
-      }
-
-      return null
-    }
-
-    return (
-      <Flex gap={1} align="center">
-        <ErrorOutlineIcon />
-        <Text muted size={1}>
-          {tooltipText()}
-        </Text>
-      </Flex>
-    )
-  }, [hasDocumentValidationErrors, isValidatingDocuments])
-
   // review screen will not be available once published
   // so redirect to summary screen
   useEffect(() => {
@@ -114,62 +75,6 @@ export const ReleaseDetail = () => {
       navigateToSummary()
     }
   }, [activeScreen, bundle?.publishedAt, navigateToSummary])
-
-  const handleConfirmPublishAll = useCallback(async () => {
-    if (!bundle) return
-
-    try {
-      setPublishBundleStatus('publishing')
-      await publishBundle(bundle._id, bundleDocuments)
-      toast.push({
-        closable: true,
-        status: 'success',
-        title: (
-          <Text muted size={1}>
-            The <strong>{bundle.title}</strong> release was published
-          </Text>
-        ),
-      })
-    } catch (publishingError) {
-      toast.push({
-        status: 'error',
-        title: (
-          <Text muted size={1}>
-            Failed to publish the <strong>{bundle.title}</strong> release
-          </Text>
-        ),
-      })
-      console.error(publishingError)
-    } finally {
-      setPublishBundleStatus('idle')
-    }
-  }, [bundle, bundleDocuments, publishBundle, toast])
-
-  const confirmPublishDialog = useMemo(() => {
-    if (publishBundleStatus === 'idle') return null
-
-    return (
-      <Dialog
-        id="confirm-publish-dialog"
-        header="Are you sure you want to publish the release and all document versions?"
-        onClose={() => setPublishBundleStatus('idle')}
-        footer={{
-          confirmButton: {
-            text: 'Publish',
-            tone: 'default',
-            onClick: handleConfirmPublishAll,
-            loading: publishBundleStatus === 'publishing',
-            disabled: publishBundleStatus === 'publishing',
-          },
-        }}
-      >
-        <Text muted size={1}>
-          The <strong>{bundle?.title}</strong> release and its {bundleDocuments.length} document
-          version{bundleDocuments.length > 1 ? 's' : ''} will be published.
-        </Text>
-      </Dialog>
-    )
-  }, [bundle?.title, bundleDocuments.length, handleConfirmPublishAll, publishBundleStatus])
 
   const header = useMemo(
     () => (
@@ -230,18 +135,12 @@ export const ReleaseDetail = () => {
           </Flex>
 
           <Flex flex="none" gap={2}>
-            {showPublishButton && (
-              <Button
-                tooltipProps={{
-                  disabled: !isPublishButtonDisabled,
-                  content: publishTooltipContent,
-                  placement: 'bottom',
-                }}
-                icon={PublishIcon}
+            {showPublishButton && bundle && (
+              <ReleasePublishAllButton
+                bundle={bundle}
+                bundleDocuments={bundleDocuments}
                 disabled={isPublishButtonDisabled}
-                text="Publish all"
-                onClick={() => setPublishBundleStatus('confirm')}
-                loading={publishBundleStatus === 'publishing'}
+                validation={validation}
               />
             )}
             <BundleMenuButton bundle={bundle} documentCount={bundleDocuments.length} />
@@ -252,15 +151,14 @@ export const ReleaseDetail = () => {
     [
       activeScreen,
       bundle,
-      bundleDocuments.length,
+      bundleDocuments,
       bundleHasDocuments,
       isPublishButtonDisabled,
       navigateToReview,
       navigateToSummary,
-      publishBundleStatus,
-      publishTooltipContent,
       router,
       showPublishButton,
+      validation,
     ],
   )
 
@@ -325,7 +223,6 @@ export const ReleaseDetail = () => {
           </Box>
         </Container>
       </Card>
-      {confirmPublishDialog}
     </Flex>
   )
 }
